Added basic material support
A "material" is a copy of the uniforms for a shader program, in order
to allow them to be changed individually for different objects using
the same shader.

Maybe in the future, will allow shader to specify if varibles are
"material level" e.g. object color, or "scene level" e.g. lighting color.